requirements.txt
streamlit
pandas
numpy
scikit-learn
transformers
torch
sentence-transformers
faiss-cpu
nltk
plotly
wordcloud
requests
beautifulsoup4